Move exec tests into a separate file
